requirements.txt
moviepy==1.0.3
Pillow==10.0.0
requests==2.31.0
openai>=1.0.0
python-dotenv==1.0.0
streamlit==1.28.0
streamlit-option-menu==0.3.6
google-generativeai>=0.3.0